Change model settings
모델 형태 변경
모델의 세부 내역 변경
모델 random seed 설정
