+++ .gitignore
@@ -1,1 +1,2 @@
 Data/*.dat
+__pycache__
